fix: Modify zip contained files for windows support
Windows does a weird thing where it zips files into a new folder named
for the archive, which is different to the behaviour on other systems.
Since windows is the most popular I will use this as the intended use
and mac systems will have to deal.

This does have an added benefit of being easier to use on SWAN, so there
is also that added benefit.